Refactoring the login page by integrating a custom hook
