+++ web/src/pages/Login.jsx
@@ -1,6 +1,4 @@
 import React, { useState } from "react";
-import { useAuth } from "../context/AuthContext"; // Contexte personnalisé pour l'authentification
-import { useNavigate } from "react-router-dom"; // Pour rediriger l'utilisateur après connexion
 
 // Import des composants UI réutilisables
 import Divider from "../components/Divider";
@@ -8,6 +6,7 @@ import SocialBtn from "../components/SocialBtn";
 import AuthForm from "../components/AuthForm";
 import AuthTitle from "../components/AuthTitle";
 import AuthState from "../components/AuthState";
+import useAuthForm from "../hooks/useAuthForm";
 
 // Import des ressources (logos et background image)
 import AppleLogo from "../assets/apple.png";
@@ -21,70 +20,16 @@ export default function Login() {
   // URL de l’API backend (défaut en local)
   const API_BASE = import.meta.env.VITE_API_BASE_URL || "http://localhost:3001";
 
-  const { login } = useAuth(); // méthode pour connecter un utilisateur via le contexte
-  const navigate = useNavigate(); // hook pour naviguer programmé
-
-  // État pour afficher les erreurs côté UI
-  const [error, setError] = useState(null);
-
   // Détermine si l'on est en mode "sign in" ou "sign up"
   const [isSignIn, setIsSignIn] = useState(true);
 
-  // Données du formulaire utilisateur
-  const [formData, setFormData] = useState({
-    name: "",
-    email: "",
-    password: "",
-  });
-
-  // Fonction à implémenter pour la connexion 
-  async function signIn(user) {
-    try {
-      const res = await fetch(`${API_BASE}/api/login`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        credentials: "include", // Pour envoyer le cookie de session
-        body: JSON.stringify(user),
-      });
-  
-      const data = await res.json();
-  
-      if (!res.ok) throw new Error(data.error || "Login failed");
-  
-      login(data.user); // Stocke l'utilisateur dans ton AuthContext
-      navigate("/");
-    } catch (error) {
-      setError(error.message);
-    }
-  }
-
-  // Fonction d'inscription
-  async function signUp(user) {
-    try {
-      const res = await fetch(`${API_BASE}/api/signup`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        credentials: "include", // Inclut les cookies, utile pour la gestion de session
-        body: JSON.stringify(user),
-      });
-
-      const data = await res.json();
-
-      // En cas d'erreur, déclenche une exception
-      if (!res.ok) throw new Error(data.error || "Unknown error");
-
-      // Sauvegarde l'utilisateur dans le contexte et redirige vers la page d’accueil
-      login(data.user);
-      navigate("/");
-    } catch (error) {
-      // Affiche l’erreur dans le composant
-      setError(error.message);
-    }
-  }
+  const {
+    formData,
+    error,
+    handleChange,
+    signUp,
+    signIn
+  } = useAuthForm(API_BASE);
 
   // Soumission du formulaire
   function handleSubmit(e) {
@@ -95,15 +40,6 @@ export default function Login() {
   // Changement entre les modes "sign in" et "sign up"
   function handleClick() {
     setIsSignIn((prev) => !prev);
-  }
-
-  // Mise à jour des données du formulaire au fur et à mesure de la saisie
-  function handleChange(event) {
-    const { name, value } = event.target;
-    setFormData((prev) => ({
-      ...prev,
-      [name]: value,
-    }));
   }
 
   return (
